+++ src/MVDegrains.h
@@ -213,14 +213,14 @@ static inline int DegrainWeight(int64_t thSAD, int64_t blockSAD) {
 }
 
 
-static inline void useBlock(const uint8_t *&p, int &np, int &WRef, int isUsable, const FakeGroupOfPlanes *fgop, int i, MVPlane * const *pPlane, const uint8_t **pSrcCur, int xx, const int *nSrcPitch, int nLogPel, int plane, int xSubUV, int ySubUV, const int *thSAD) {
+static inline void useBlock(const uint8_t *&p, int &np, int &WRef, int isUsable, const FakeGroupOfPlanes *fgop, int i, MVPlane * const *pPlane, const uint8_t **pSrcCur, int xx, const int *nSrcPitch, int nLogPel, int plane, int xSubUV, int ySubUV, const int64_t *thSAD) {
     if (isUsable) {
         const FakeBlockData *block = fgopGetBlock(fgop, 0, i);
         int blx = (block->x << nLogPel) + block->vector.x;
         int bly = (block->y << nLogPel) + block->vector.y;
         p = mvpGetPointer(pPlane[plane], plane ? blx >> xSubUV : blx, plane ? bly >> ySubUV : bly);
         np = pPlane[plane]->nPitch;
-        int blockSAD = block->vector.sad;
+        int64_t blockSAD = block->vector.sad;
         WRef = DegrainWeight(thSAD[plane], blockSAD);
     } else {
         p = pSrcCur[plane] + xx;
